# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Hotel Reservation System API",
        "description": "API REST para el sistema de reservas de hotel. Incluye gestión de amenities, tipos de habitación y habitaciones físicas.",
        "contact": {
            "name": "API Support",
            "email": "adiazy@uamv.edu.ni"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/rate-plan-prices": {
            "get": {
                "tags": [
                    "Rate Plan Prices"
                ],
                "summary": "Obtener lista de precios de planes de tarifa",
                "description": "Lista de precios de planes de tarifa\n\nFiltros opcionales vía query params:\n- ?rate_plan_id=1\n- ?room_type_id=2\n- ?from=2025-01-01\n- ?to=2025-01-31",
                "operationId": "431c4473ab52db93d3a36a12ad77ffbd",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número de página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "rate_plan_id",
                        "in": "query",
                        "description": "Filtrar por plan de tarifa",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "room_type_id",
                        "in": "query",
                        "description": "Filtrar por tipo de habitación",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 2
                        }
                    },
                    {
                        "name": "from",
                        "in": "query",
                        "description": "Fecha desde (formato: YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2025-01-01"
                        }
                    },
                    {
                        "name": "to",
                        "in": "query",
                        "description": "Fecha hasta (formato: YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2025-01-31"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de precios obtenida exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Rate Plan Prices"
                ],
                "summary": "Crear un nuevo precio de plan de tarifa",
                "description": "Crear un nuevo precio de plan de tarifa",
                "operationId": "7e1850cf46ddfd795478ab3523234dec",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "rate_plan_id",
                                    "room_type_id",
                                    "date_from",
                                    "date_to",
                                    "occupancy",
                                    "price"
                                ],
                                "properties": {
                                    "rate_plan_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "room_type_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "date_from": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-01-01"
                                    },
                                    "date_to": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-01-31"
                                    },
                                    "occupancy": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 80
                                    },
                                    "extra_adult": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 10,
                                        "nullable": true
                                    },
                                    "extra_child": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 5,
                                        "nullable": true
                                    },
                                    "currency": {
                                        "type": "string",
                                        "maxLength": 3,
                                        "example": "NIO",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Precio creado exitosamente"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            }
        },
        "/api/rate-plan-prices/{id}": {
            "get": {
                "tags": [
                    "Rate Plan Prices"
                ],
                "summary": "Obtener un precio de plan de tarifa por ID",
                "description": "Obtener un precio de plan de tarifa específico",
                "operationId": "9bf90486960423250a658804f665c977",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del precio",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Precio obtenido exitosamente"
                    },
                    "404": {
                        "description": "Precio no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Rate Plan Prices"
                ],
                "summary": "Actualizar un precio de plan de tarifa",
                "description": "Actualizar un precio de plan de tarifa",
                "operationId": "6a6a4788f568e504fedd6ba6582b7a5b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del precio",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "rate_plan_id": {
                                        "type": "integer"
                                    },
                                    "room_type_id": {
                                        "type": "integer"
                                    },
                                    "date_from": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "date_to": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "occupancy": {
                                        "type": "integer"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "extra_adult": {
                                        "type": "number",
                                        "format": "float",
                                        "nullable": true
                                    },
                                    "extra_child": {
                                        "type": "number",
                                        "format": "float",
                                        "nullable": true
                                    },
                                    "currency": {
                                        "type": "string",
                                        "maxLength": 3,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Precio actualizado exitosamente"
                    },
                    "404": {
                        "description": "Precio no encontrado"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Rate Plan Prices"
                ],
                "summary": "Eliminar un precio de plan de tarifa",
                "description": "Eliminar un precio de plan de tarifa",
                "operationId": "5020e9b8196ba75be2ddb6d92952ceab",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del precio",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Precio eliminado exitosamente"
                    },
                    "404": {
                        "description": "Precio no encontrado"
                    }
                }
            }
        },
        "/api/rate-plan-prices/lookup": {
            "get": {
                "tags": [
                    "Rate Plan Prices"
                ],
                "summary": "Buscar precio para una fecha específica",
                "description": "Lookup de precio para una fecha específica\n\nBusca el primer registro donde:\n- rate_plan_id = ?\n- room_type_id = ?\n- occupancy = ?\n- date_from <= date <= date_to",
                "operationId": "ffd1340eb3676d4de2cf3a6fa0ed7baa",
                "parameters": [
                    {
                        "name": "rate_plan_id",
                        "in": "query",
                        "description": "ID del plan de tarifa",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "room_type_id",
                        "in": "query",
                        "description": "ID del tipo de habitación",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 2
                        }
                    },
                    {
                        "name": "date",
                        "in": "query",
                        "description": "Fecha a consultar (formato: YYYY-MM-DD)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2025-01-10"
                        }
                    },
                    {
                        "name": "occupancy",
                        "in": "query",
                        "description": "Ocupación (número de huéspedes)",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 2
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Precio encontrado exitosamente"
                    },
                    "404": {
                        "description": "No se encontró precio para los criterios especificados"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            }
        },
...
    },
...
}